refactor: centralize dayjs imports and remove legacy dayjs utility module
