+++ src/api/gtfsService.ts
@@ -5,7 +5,7 @@ import {
   GtfsRideStopPydanticModel,
   GtfsRideWithRelatedPydanticModel,
 } from 'open-bus-stride-client'
-import dayjs from '../pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 import { API_CONFIG, BASE_PATH, MAX_HITS_COUNT } from 'src/api/apiConfig'
 import { BusRoute, fromGtfsRoute } from 'src/model/busRoute'
 import { BusStop, fromGtfsStop } from 'src/model/busStop'
