+++ src/hooks/useSingleVehicleData.ts
@@ -2,7 +2,7 @@ import { useContext, useEffect, useMemo, useState } from 'react'
 import useVehicleLocations from 'src/api/useVehicleLocations'
 import { SearchContext } from 'src/model/pageState'
 import { Point } from 'src/pages/timeBasedMap'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 
 export const useSingleVehicleData = (vehicleRef?: number) => {
   const {
