+++ src/pages/gapsPatterns/useGapsList.ts
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react'
 import { getGapsAsync } from '../../api/gapsService'
 import { HourlyData, sortByMode } from '../components/utils'
 import { GapsList } from 'src/model/gaps'
-import dayjs from 'src/pages/components/utils/dayjs'
+import dayjs from 'src/dayjs'
 
 type HourlyDataList = HourlyData[]
 // Convert gapsList into HourlyDataList structure
